archive/6-solutions/common-infra/ANF-S3-MCP/server-s3mcp/pyproject.toml
[project]
name = "s3-mcp-server"
version = "1.0.0"
description = "A comprehensive MCP server for S3 integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "S3 MCP Server Contributors"}
]
keywords = ["s3", "mcp", "aws", "api", "server"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
"fastmcp>=2.8.1",
"boto3>=1.28.0",
"python-dotenv>=1.0.0"
]
requires-python = ">=3.10"

[project.urls]
Homepage = "https://github.com/konstantinasm/s3-mcp"
Repository = "https://github.com/konstantinasm/s3-mcp"
Issues = "https://github.com/konstantinasm/s3-mcp/issues"
Documentation = "https://github.com/konstantinasm/s3-mcp#readme"

[project.scripts]
s3-mcp = "src.s3_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
